requirements.txt
absl-py==0.11.0
astunparse==1.6.3
azure-ai-formrecognizer==3.1.0b3
azure-common==1.1.26
azure-core==1.13.0b1
breadability==0.1.20
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cmake==3.18.4.post1
cycler==0.10.0
dataclasses==0.8
dateparser==1.0.0
decorator==4.4.2
dlib==19.21.1
docopt==0.6.2
dominant-color-detection==0.1.8
easyocr==1.2.3
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gTTS==2.2.1
h5py==2.10.0
idna==2.10
imageai==2.1.5
imageio==2.9.0
importlib-metadata==3.4.0
imutils==0.5.4
invoice2data==0.3.5
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
jsonpickle==2.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
msrest==0.6.21
nest-asyncio==1.3.2
networkx==2.5
nltk==3.5
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pdf2text==1.0.0
pdfminer==20191125
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycryptodome==3.10.1
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
pyttsx3==2.71
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.16.2
scikit-learn==0.24.1
scipy==1.5.4
serial==0.0.97
six==1.15.0
SpeechRecognition==3.8.1
sumy==0.8.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
termcolor==1.1.0
tf-slim==1.1.0
tflite-runtime==2.5.0
thinning-py3==1.2.3
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.1
torchvision==0.5.0
tqdm==4.56.0
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.2
webcolors==1.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
